Fix bug where fixed length buffer spans can be corrupted
It turns out that what I had before *appeared* to work most of the time, but as I was refactoring, I came across some bizarre corruption that can only be explained by memory moving around. Taking unsafe spans without informing the compiler of this is evidently a bad idea.

Appling `[UnscopedRef]` to span-returning properties in structs is one critical step to the fix.
This commit applies this change.

The other requirement is for properties on *classes* that return these struct-based spans to use `ref return` so that the reference remains pinned to their respective callers.
This change will be in a follow-up commit.