Refactor ZcashWallet to use a flat list of accounts
HDWallet no longer contains the accounts based on it. Having a tree of collections of accounts was just too complicated.

This commit also switches the means of de-duping objects in the msgpack file to one handled at the resolver level. This makes things much simpler and solves the de-duping problem for all reference types.