Add rust calculation of meaningful user balances
This new struct describes balances in a way that a user is likely to (need to) understand. Some fields may be too technical and may be omitted by an app or combined with other fields.

Balances are divided such that an app that chooses a different policy for what to convey to the user should be able to add or subtract fields from each other to produce whatever value they see fit.