Log the mnemonic that a test randomly generates
This is because in one particular build on a GitHub Actions mac agent, somehow the random generation seems to have led to the generation of an orchard address that collided with a one that is hard-coded in the test. This seems unbelievable. So if it ever happens again, I want to be able to reproduce it.