Support for multiple receiving addresses to be stored per contact
This is something of a prerequisite for displaying names in the History view, because users should be able to type a name into blank cells and have that transaction be linked to the sender or recipient of the transaction, based on the address, even if the named contact already has a (different) receiving address listed.
Also print the contact *name* in the Send view's To address when the send is initiated from the address book.