`DefaultAddress` UAs should use a consistent diversifier index
As part of this, we add an API to make it easy to construct a `UnifiedAddress` using keys instead of addresses, so that all the indexes match.

This improves ZIP-316 compliance, which for contestable reasons [requires all receivers in a UA to have matching indexes][ZIP316].

Note that `UnifiedAddress.Create` can still combine arbitrary addresses. It's only the `TryCreate` method that guarantees matching indexes in the result. But since `TryCreate` is now how `UnifiedViewingKey.DefaultAddress`  is implemented, default addresses will now conform to ZIP-316.
Note also that `ZcashAccount.GetDiversifiedAddress` already generated compliant UAs (and always lack a transparent component).

[ZIP316]: https://github.com/zcash/zips/blob/fee271c03da36836c58e649d9674fa430e343810/zip-0316.rst#L690-L692