Ship the desktop app as self-contained on Windows too
Previous to this, we built a self-contained app only for mac and linux. This left Windows as the only platform that required elevated privileges to install (in the common case that the user doesn't have .NET 8 installed already). In user testing, the UAC prompt on Windows came up for every user and confused every user.