Fix unstable `ZcashAddress.IsMatch_PartialMatch` test
This was caused by missing explicit `Equals` methods on the `OrchardReceiver` struct, leading to the default `ValueType.Equals` behavior that is inadequate.
I made some other changes to help ensure the fix is and remains complete.

I also added the `TexReceiver` to the `IsMatch` method since it was missing there.