Build an iOS framework for Nerdbank.Zcash
This is required so that consuming apps will not be rejected by the Apple App Store.

Fix target platform versions

TPV to 17.0

Target ios17.5 instead of ios18.0

This is because ios18.0 requires Xcode 16, but Visual Studio only knows how to work with Xcode 15.4. Thus ios17.5 is the latest version that we can build.

Add missing manifest

Log when generating C# bindings

Simplify bindings code

Drop the TPV on the app

Fix minimum OS version